Move all public headers to the same location.
